bitcoin: split xpub and transaction signing to separate screens
Also improve copywriting from design feedback, and sign transaction with
explicit button.
