+++ src/lib/routeNames.js
@@ -27,4 +27,7 @@ export const ROUTE_NAMES = {
   URBIT_ID: Symbol('URBIT_ID'),
   RESIDENTS: Symbol('RESIDENTS'),
   ADOPT: Symbol('ADOPT'),
+  BITCOIN: Symbol('BITCOIN'),
+  BITCOIN_XPUB: Symbol('BITCOIN_XPUB'),
+  BITCOIN_SIGN_TRANSACTION: Symbol('BITCOIN_SIGN_TRANSACTION'),
 };
